migration_tools/requirements.txt
libclang>=18.1.1
mcp>=1.1.2
anyio>=4.2.0
Wand
Pillow
pygltflib>=1.16.3
numpy